Docker dependency refactor
