+++ iDiff/utils/image_utils.go
@@ -2,44 +2,15 @@ package utils
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
 	"io"
 	"os"
 	"path/filepath"
 	"regexp"
-	"strings"
 
-	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/client"
 	"github.com/docker/docker/pkg/system"
 	"github.com/golang/glog"
 )
-
-func saveImageToTar(image string) (string, error) {
-	cli, err := client.NewEnvClient()
-	if err != nil {
-		return "", err
-	}
-
-	fromImage := image
-	toTar := image
-	// If not an already existing image ID, have to pull it from a repo before saving it
-	if !CheckImageID(image) {
-		imageID, imageName, err := pullImageFromRepo(cli, image)
-		if err != nil {
-			return "", err
-		}
-		fromImage = imageID
-		toTar = imageName
-	}
-	// Convert the image into a tar
-	imageTarPath, err := ImageToTar(cli, fromImage, toTar)
-	if err != nil {
-		return "", err
-	}
-	return imageTarPath, nil
-}
 
 // ImageToDir converts an image to an unpacked tar and creates a representation of that directory.
 func ImageToDir(img string, eng bool) (string, string, error) {
@@ -70,72 +41,6 @@ func ImageToDir(img string, eng bool) (string, string, error) {
 	return TarToJSON(tarName)
 }
 
-type Event struct {
-	Status         string `json:"status"`
-	Error          string `json:"error"`
-	Progress       string `json:"progress"`
-	ProgressDetail struct {
-		Current int `json:"current"`
-		Total   int `json:"total"`
-	} `json:"progressDetail"`
-}
-
-func getImagePullResponse(image string, response []Event) (string, error) {
-	var imageDigest string
-	for _, event := range response {
-		if event.Error != "" {
-			err := fmt.Errorf("Error pulling image %s: %s", image, event.Error)
-			return "", err
-		}
-		digestPattern := regexp.MustCompile("^Digest: (sha256:[a-z|0-9]{64})$")
-		digestMatch := digestPattern.FindStringSubmatch(event.Status)
-		if len(digestMatch) != 0 {
-			imageDigest = digestMatch[1]
-			return imageDigest, nil
-		}
-	}
-	err := fmt.Errorf("Could not pull image %s", image)
-	return "", err
-}
-
-func processImagePullEvents(image string, events []Event) (string, string, error) {
-	imageDigest, err := getImagePullResponse(image, events)
-	if err != nil {
-		return "", "", err
-	}
-
-	URLPattern := regexp.MustCompile("^.+/(.+(:.+){0,1})$")
-	URLMatch := URLPattern.FindStringSubmatch(image)
-	imageName := strings.Replace(URLMatch[1], ":", "", -1)
-	imageURL := strings.TrimSuffix(image, URLMatch[2])
-	imageID := imageURL + "@" + imageDigest
-
-	return imageID, imageName, nil
-}
-
-func pullImageFromRepo(cli client.APIClient, image string) (string, string, error) {
-	response, err := cli.ImagePull(context.Background(), image, types.ImagePullOptions{})
-	if err != nil {
-		return "", "", err
-	}
-	defer response.Close()
-
-	d := json.NewDecoder(response)
-
-	var events []Event
-	for {
-		var event Event
-		if err := d.Decode(&event); err != nil {
-			if err == io.EOF {
-				break
-			}
-			return "", "", err
-		}
-		events = append(events, event)
-	}
-	return processImagePullEvents(image, events)
-}
-
 // ImageToTar writes an image to a .tar file
 func ImageToTar(cli client.APIClient, image, tarName string) (string, error) {
 	imgBytes, err := cli.ImageSave(context.Background(), []string{image})
@@ -145,6 +50,47 @@ func ImageToTar(cli client.APIClient, image, tarName string) (string, error) {
 	defer imgBytes.Close()
 	newpath := tarName + ".tar"
 	return newpath, copyToFile(newpath, imgBytes)
+}
+
+func CheckImageID(image string) bool {
+	pattern := regexp.MustCompile("[a-z|0-9]{12}")
+	if exp := pattern.FindString(image); exp != image {
+		return false
+	}
+	return true
+}
+
+func CheckImageURL(image string) bool {
+	pattern := regexp.MustCompile("^.+/.+(:.+){0,1}$")
+	if exp := pattern.FindString(image); exp != image || CheckTar(image) {
+		return false
+	}
+	return true
+}
+
+func saveImageToTar(image string) (string, error) {
+	cli, err := client.NewEnvClient()
+	if err != nil {
+		return "", err
+	}
+
+	fromImage := image
+	toTar := image
+	// If not an already existing image ID, have to pull it from a repo before saving it
+	if !CheckImageID(image) {
+		imageID, imageName, err := pullImageFromRepo(cli, image)
+		if err != nil {
+			return "", err
+		}
+		fromImage = imageID
+		toTar = imageName
+	}
+	// Convert the image into a tar
+	imageTarPath, err := ImageToTar(cli, fromImage, toTar)
+	if err != nil {
+		return "", err
+	}
+	return imageTarPath, nil
 }
 
 // copyToFile writes the content of the reader to the specified file
@@ -172,20 +118,4 @@ func copyToFile(outfile string, r io.Reader) error {
 	}
 
 	return nil
-}
-
-func CheckImageID(image string) bool {
-	pattern := regexp.MustCompile("[a-z|0-9]{12}")
-	if exp := pattern.FindString(image); exp != image {
-		return false
-	}
-	return true
-}
-
-func CheckImageURL(image string) bool {
-	pattern := regexp.MustCompile("^.+/.+(:.+){0,1}$")
-	if exp := pattern.FindString(image); exp != image || CheckTar(image) {
-		return false
-	}
-	return true
 }
